New structure
Made a new structure as outlined in the diagram I made.

The `manager.js` now manages the commands and sends only one block at a
time.

The `generator.js` is a thin and dumb wrapper that simply pipes
`generator.py`'s output to the Writable Stream object that is passed.
Its most important function is calling `.end()` on the Writable Stream
when '!ENDSTREAM!' is received on stderr.

This setup is probably the fastest possible, as the data is piped
directly to the recipient via the Writable Stream response object,
coming from the Readable Stream stdout of `generator.py`.