Basic startup/shutdown working. Hangs on msg.
The `generator.js` spec shows that it starts and shuts down nicely, the
`generator.py` process throws errors as should be.

It does seem to hang when a message is passed. Don't know why this is.
The python process does thrown an error if it's thrown before listening
to stdin, but not after the blocking stdin function. Is the py process
even still alive?