Manager.js test in place
The basic functionality is there. To be able to easily use request and
response streams, `managerTest.js` uses a server. The basic tests are
passing.

There's a problem with premature ending of the WritableStream passed to
generator.js: it seems that `!ENDSTREAM!` gets sent before `stdout` has
been completely emptied, thereby prematurely ending the stream. So this
has to be accounted for, either at the Node side or the Python side.