Using stdio[3] instead of stdout
To circumvent the problems with modules printing to stdout, I'm opening
and using stdio[3] instead. Note: on the python side, when doing:
`fdopen( 3, 'w', BUFSIZE )`,
BUFSIZE must be 0 (unbuffered), this seems to work best. If not
specified, the standard bufsize is used, which is apparently pretty
small and cuts off the output.

When piping `generator.py` `stdio[3]`, `{end: false}` must be set,
otherwise, when an error occurs, the recipient will be ended before
'generr' can be emitted.

`managerTest.js` now contains a proper test for the tessellation.